+++ NEWS.md
@@ -1,3 +1,6 @@
+# rethnicity (0.1.1)
+* fixed the include path problem and suppressing warning problem, as requested and resubmit
+
 # rethnicity (0.1.0)
 * first submission of the package
 * provide `predict_ethnicity()` function to predict the ethnicity from names